folder slice

+++ src/Component/DisplayFiles.tsx
@@ -1,21 +1,32 @@
-import React, {FC, useEffect, useState} from "react";
-import {emptyDirectusFileType} from "../type/ModifiedFileType";
+import {useEffect, useState} from "react";
 import folder from "../lib/folder";
 import forum from "../lib/forum";
-import {emptyFolderType} from "../type/DirectusType";
-import {FolderType} from "@directus/sdk";
+import {FileType, FolderType} from "@directus/sdk";
 import LoadingSpinner from "./LoadingSpinner";
 import {createPortal} from "react-dom";
+import { useAppDispatch, useAppSelector } from "../App/hooks";
+import { fetchFolderById, fetchFolderByParent, setActualFolder } from "../slicers/folder-slice";
 
-const DisplayFiles: FC<{ callbackOnClick?: Function, setFolderId?: Function, showDelete?: boolean, startingFolder?: string }> = ({callbackOnClick, setFolderId, showDelete, startingFolder}) => {
-    const [files, setFiles] = useState(new Array(emptyDirectusFileType));
-    const [folders, setFolders] = useState(new Array(emptyFolderType));
+type Props = {
+    callbackOnClick?: Function;
+    showDelete?: boolean;
+    startingFolderId?: string;
+  };
+
+export default function DisplayFiles (props: Props) {
+    const { callbackOnClick, showDelete, startingFolderId } = props;
+    
     const rootFolder = {id: "", name: "Root", parent: ""} as FolderType;
-    const [actualFolder, setActualFolder] = useState(rootFolder);
+    const [files, setFiles] = useState<FileType[]>([]);
+    // const [folders, setFolders] = useState<FolderType[]>([]);
+    // const [actualFolder, setActualFolder] = useState<FolderType>(rootFolder);
     const [isLoading, setIsLoading] = useState(false);
-
     const [showPopup, setShowPopup] = useState(false);
-    const [tmpFile, setTmpFile] = useState(emptyDirectusFileType);
+    const [tmpFile, setTmpFile] = useState<FileType | null>(null);
+
+    const dispatch = useAppDispatch();
+    const { actualFolder, folderList } = useAppSelector(state => state.folder);
+
     // const [showPopup2, setShowPopup2] = useState(false);
     // const [tmpFolder, setTmpFolder] = useState(emptyFolderType);
 
@@ -23,13 +34,11 @@ const DisplayFiles: FC<{ callbackOnClick?: Function, setFolderId?: Function, sho
         setIsLoading(true);
         forum.connection().then(async () => {
             let parentId = null;
-            if (startingFolder) {
-                const folderReponse = await folder.getFolderById(startingFolder);
-                parentId = folderReponse.id;
-                setActualFolder(folderReponse);
+            if (startingFolderId) {
+                dispatch(fetchFolderById(startingFolderId));
+                parentId = actualFolder.id;
             }
-            const foldersResponse = await folder.getFolderByParent(parentId);
-            setFolders(foldersResponse.data);
+            dispatch(fetchFolderByParent(parentId));
             const filesResponse = await folder.getFilesByFolder(parentId);
             setFiles(filesResponse.data);
             setIsLoading(false);
@@ -38,29 +47,23 @@ const DisplayFiles: FC<{ callbackOnClick?: Function, setFolderId?: Function, sho
 
     async function handleClickFolder(newFolder: FolderType) {
         setIsLoading(true);
-        let foldersResponse = await folder.getFolderByParent(newFolder.id)
-        setFolders(foldersResponse.data);
+        dispatch(fetchFolderByParent(newFolder.id));
         let filesResponse = await folder.getFilesByFolder(newFolder.id);
         setFiles(filesResponse.data);
-        setActualFolder(newFolder);
-        if (setFolderId) setFolderId(newFolder.id);
+        dispatch(setActualFolder(newFolder));
         setIsLoading(false);
     }
 
     async function handleClickBack() {
         if (actualFolder.parent === '') return;
         setIsLoading(true);
-        let foldersResponse = await folder.getFolderByParent(actualFolder.parent);
-        setFolders(foldersResponse.data);
+        dispatch(fetchFolderByParent(actualFolder.parent));
         let filesResponse = await folder.getFilesByFolder(actualFolder.parent);
         setFiles(filesResponse.data);
         if (actualFolder.parent === null) {
             setActualFolder(rootFolder);
-            if (setFolderId) setFolderId('');
         } else {
-            let folderResponse = await folder.getFolderById(actualFolder.parent);
-            setActualFolder(folderResponse);
-            if (setFolderId) setFolderId(folderResponse.id);
+            dispatch(fetchFolderById(actualFolder.parent));
         }
         setIsLoading(false);
     }
@@ -70,6 +73,7 @@ const DisplayFiles: FC<{ callbackOnClick?: Function, setFolderId?: Function, sho
     }
 
     async function deleteFile() {
+        if (!tmpFile) return;
         await folder.deleteFile(tmpFile);
         window.location.reload();
     }
@@ -83,9 +87,9 @@ const DisplayFiles: FC<{ callbackOnClick?: Function, setFolderId?: Function, sho
                             Dossier actuel : {actualFolder.name}
                         </div>
                         {
-                            folders && (
+                            folderList && (
                                 <div>
-                                    {folders.map((folder) => {
+                                    {folderList.map((folder) => {
                                         return (
                                             <>
                                                 <div onClick={() => handleClickFolder(folder)} className={"cursor-pointer"}>
@@ -133,7 +137,7 @@ const DisplayFiles: FC<{ callbackOnClick?: Function, setFolderId?: Function, sho
                     <div className={"alertContainer"}>
                         <div className={"alertPopup text-center"}>
                             <h1>Êtes-vous sur de vouloir supprimer ce fichier ?</h1>
-                            <h3>"{tmpFile.filename_download}"</h3>
+                            <h3>"{tmpFile!.filename_download}"</h3>
                             <div className={"flex justify-evenly"}>
                                 <button className={"bg-green-500 hover:bg-green-700 text-white w-1/6 font-bold py-2 px-4 rounded"} onClick={deleteFile}>Oui</button>
                                 <button className={"bg-red-500 hover:bg-red-700 text-white w-1/6 font-bold py-2 px-4 rounded"} onClick={quitPopup}>Annuler</button>
@@ -145,5 +149,3 @@ const DisplayFiles: FC<{ callbackOnClick?: Function, setFolderId?: Function, sho
         </>
     )
 }
-
-export default DisplayFiles;
